move enums to enums package and some structural changes to the endpoint and transfer types
