+++ backend/src/main/kotlin/dat257/gyro/backend/database/models/WalkHistory.kt
@@ -1,7 +1,7 @@
 package dat257.gyro.backend.database.models
 
 import dat257.gyro.backend.database.tables.WalkHistories
-import dat257.gyro.shared.types.Rating
+import dat257.gyro.shared.enums.Rating
 import org.jetbrains.exposed.dao.IntEntity
 import org.jetbrains.exposed.dao.IntEntityClass
 import org.jetbrains.exposed.dao.id.EntityID
