+++ shared/src/main/kotlin/dat257/gyro/shared/ApiHelper.kt
@@ -3,7 +3,7 @@ package dat257.gyro.shared
 class ApiHelper {
     companion object {
         fun getApiUrl(endpoint: String? = null): String {
-            var url = System.getenv("API_ENDPOINT_URL") ?: "localhost:9000"
+            var url = System.getenv("API_ENDPOINT_URL") ?: "http://localhost:9000"
             if (endpoint != null) {
                 if (endpoint[0] != '/') {
                     url += "/"
