--- shared/src/main/kotlin/dat257/gyro/shared/StatusCodes.kt
@@ -1,13 +1,0 @@
-package dat257.gyro.shared
-
-enum class StatusCode(code: Int) {
-    OK(200),
-    BadRequest(400),
-    Unauthorized(401),
-    Forbidden(403),
-    NotFound(404),
-    MethodNotAllowed(405),
-    InternalServerError(500),
-    NotImplemented(501),
-    Unknown(0)
-}
