+++ shared/src/main/kotlin/dat257/gyro/shared/enums/Rating.kt
@@ -1,4 +1,4 @@
-package dat257.gyro.shared.types
+package dat257.gyro.shared.enums
 
 enum class Rating(val value: Int) {
     ZERO(0),
