+++ shared/src/main/kotlin/dat257/gyro/shared/enums/ShowOnMap.kt
@@ -1,4 +1,4 @@
-package dat257.gyro.shared.types
+package dat257.gyro.shared.enums
 
 enum class ShowOnMap(val value: String) {
     YES("yes"),
