+++ shared/src/main/kotlin/dat257/gyro/shared/types/EndpointType.kt
@@ -1,15 +1,15 @@
 package dat257.gyro.shared.types
 
-import dat257.gyro.shared.StatusCode
 import org.http4k.client.JavaHttpClient
 import org.http4k.core.HttpHandler
+import org.http4k.core.Status
 
-abstract class BaseEndpoint<DataType : TransferType<DataType>> {
+abstract class EndpointType<DataType : TransferType<DataType>> {
     protected abstract val endpointUrl: String
     protected abstract val data: DataType
     protected val client: HttpHandler = JavaHttpClient()
 
-    abstract fun store(): StatusCode
-    abstract fun get(): Pair<StatusCode, DataType>
-    abstract fun delete(): StatusCode
+    abstract fun store(): Status
+    abstract fun get(): Pair<Status, DataType>
+    abstract fun delete(): Status
 }
